Initial Release of Channel Overclocker Feature
Quite alot of work for something so "simple". This adds the Item
"Channel Overclocker" which allows cards to have up to 256 Channels
(0-255). The Slot for it only appears when you have the upgrade in your
Inventory. Additionally cards now update their channel colors on change.
When more than 16 Channels are used the colored part is split in half.
More design choices are:
- The upgrade shows its channel color once put into its Slot (it looks
nicer in JEI without it)
- Added shapeless recipe to clear channel on upgrade
- Once added, a number button appears to provide easier channel change

TODO: GUI for Channel Overclocker (when Right-clicked, allow for stacked edits)